2.0.0.2 git-svn-id: https://casparcg.svn.sourceforge.net/svnroot/casparcg/server/branches/2.0.0.2@322 362d55ac-95cf-4e76-9f9a-cbaa9c17b72d

+++ core/producer/flash/FlashAxContainer.h
@@ -232,7 +232,7 @@ public:
 	STDMETHOD(Unadvise)(DWORD dwCookie);
 	STDMETHOD(Freeze)(BOOL fFreeze);
 	STDMETHOD(GetTime)(VARIANT *pvtime);
-	int GetFPS();
+	double GetFPS();
 
 	HRESULT CreateAxControl();
 	void DestroyAxControl();
